feat: some optimizations
