+++ packages/verified-fetch/src/utils/enhanced-dag-traversal.ts
@@ -1,15 +1,15 @@
 import { type ComponentLogger } from '@libp2p/interface'
 import { type ReadableStorage, exporter, type ExporterOptions } from 'ipfs-unixfs-exporter'
 import first from 'it-first'
+// import peekable from 'it-peekable'
 import toBrowserReadableStream from 'it-to-browser-readablestream'
 import { type CID } from 'multiformats/cid'
 import { type ContentTypeParser } from '../types.js'
-import { setContentType } from './set-content-type.js'
+import { getContentType } from './get-content-type.js'
 
 export interface EnhancedDagTraversalOptions extends ExporterOptions {
   blockstore: ReadableStorage
   cidOrPath: string | CID
-  response: Response
   logger: ComponentLogger
   path: string
   contentTypeParser?: ContentTypeParser
@@ -29,12 +29,10 @@ export async function enhancedDagTraversal ({
   length,
   path,
   logger,
-  contentTypeParser,
-  response
+  contentTypeParser
 }: EnhancedDagTraversalOptions): Promise<EnhancedDagTraversalResponse> {
   const log = logger.forComponent('helia:verified-fetch:enhanced-dag-traversal')
 
-  // Fetch the first chunk eagerly
   const dfsEntry = await exporter(cidOrPath, blockstore, {
     signal,
     onProgress,
@@ -48,9 +46,10 @@ export async function enhancedDagTraversal ({
     length
   })
 
-  let firstChunk
+  let firstChunk: Uint8Array = new Uint8Array()
   let error: Error
   try {
+    // Fetch the first chunk eagerly
     firstChunk = await first(dfsIter)
   } catch (err: any) {
     if (signal?.aborted === true) {
@@ -61,43 +60,38 @@ export async function enhancedDagTraversal ({
     }
   }
 
-  // Determine content type based on the first chunk
-  await setContentType({ bytes: firstChunk, path, response, contentTypeParser, log })
-
-  const contentType = response.headers.get('content-type')
-  const isImageOrVideo = contentType?.startsWith('video/') === true || contentType?.startsWith('image/') === true
-  log.trace('Content type determined: %s', contentType)
-
-  const enhancedIter = async function * (): AsyncGenerator<Uint8Array, void, undefined> {
-    if (error != null) {
-      throw error
-    }
-    if (isImageOrVideo) {
-      yield * dfsIter
-      return
-    }
-
-    // If not image/video, switch to a BFS iterator
-    const bfsEntry = await exporter(cidOrPath, blockstore, {
-      signal,
-      onProgress
-    })
-
-    const bfsIter = bfsEntry.content({
-      signal,
-      onProgress,
-      offset,
-      length
-    })
-
-    // continue with the BFS iterator
-    yield * bfsIter
-  }
-
-  const stream = toBrowserReadableStream(enhancedIter())
-
   return {
-    stream,
+    stream: toBrowserReadableStream({
+      [Symbol.asyncIterator]: async function * (): AsyncGenerator<Uint8Array, void, undefined> {
+        if (error != null) {
+          throw error
+        }
+
+        // Determine content type based on the first chunk
+        const contentType = await getContentType({ bytes: firstChunk, contentTypeParser, path, log })
+
+        const isImageOrVideo = contentType.startsWith('video/') || contentType.startsWith('image/')
+        log.trace('Content type determined: %s', contentType)
+
+        const exporterEntry = isImageOrVideo
+          ? dfsEntry
+        // If not image/video, switch to a BFS iterator
+          : await exporter(cidOrPath, blockstore, {
+            signal,
+            onProgress
+          })
+
+        // continue with the BFS iterator
+        for await (const chunk of exporterEntry.content({
+          signal,
+          onProgress,
+          offset,
+          length
+        })) {
+          yield chunk
+        }
+      }
+    }),
     firstChunk
   }
 }
