+++ packages/verified-fetch/src/utils/set-content-type.ts
@@ -1,6 +1,6 @@
 import { type Logger } from '@libp2p/interface'
 import { type ContentTypeParser } from '../types.js'
-import { isPromise } from './type-guards.js'
+import { getContentType } from './get-content-type.js'
 
 export interface SetContentTypeOptions {
   bytes: Uint8Array
@@ -12,27 +12,7 @@ export interface SetContentTypeOptions {
 }
 
 export async function setContentType ({ bytes, path, response, contentTypeParser, log, defaultContentType = 'application/octet-stream' }: SetContentTypeOptions): Promise<void> {
-  let contentType: string | undefined
-
-  if (contentTypeParser != null) {
-    try {
-      let fileName = path.split('/').pop()?.trim()
-      fileName = fileName === '' ? undefined : fileName
-      const parsed = contentTypeParser(bytes, fileName)
-
-      if (isPromise(parsed)) {
-        const result = await parsed
-
-        if (result != null) {
-          contentType = result
-        }
-      } else if (parsed != null) {
-        contentType = parsed
-      }
-    } catch (err) {
-      log.error('error parsing content type', err)
-    }
-  }
+  const contentType = await getContentType({ bytes, contentTypeParser, path, log, defaultContentType })
   log.trace('setting content type to "%s"', contentType ?? defaultContentType)
   response.headers.set('content-type', contentType ?? defaultContentType)
 }
