+++ packages/verified-fetch/src/verified-fetch.ts
@@ -24,7 +24,6 @@ import { getContentDispositionFilename } from './utils/get-content-disposition-f
 import { getETag } from './utils/get-e-tag.js'
 import { getPeerIdFromString } from './utils/get-peer-id-from-string.js'
 import { getResolvedAcceptHeader } from './utils/get-resolved-accept-header.js'
-// import { getStreamFromAsyncIterable } from './utils/get-stream-from-async-iterable.js'
 import { tarStream } from './utils/get-tar-stream.js'
 import { getRedirectResponse } from './utils/handle-redirects.js'
 import { parseResource } from './utils/parse-resource.js'
@@ -376,25 +375,7 @@ export class VerifiedFetch {
     this.log.trace('calling exporter for %c/%s with offset=%o & length=%o', resolvedCID, path, offset, length)
 
     try {
-      // const entry = await exporter(resolvedCID, this.helia.blockstore, {
-      //   signal: options?.signal,
-      //   onProgress: options?.onProgress
-      // })
-
-      // const asyncIter = entry.content({
-      //   signal: options?.signal,
-      //   onProgress: options?.onProgress,
-      //   offset,
-      //   length
-      // })
-      this.log('got async iterator for %c/%s', cid, path)
-
-      // const { stream, firstChunk } = await getStreamFromAsyncIterable(asyncIter, path ?? '', this.helia.logger, {
-      //   onProgress: options?.onProgress,
-      //   signal: options?.signal
-      // })
-      const tmpResponse = new Response()
-      const { stream } = await enhancedDagTraversal({
+      const { firstChunk, stream } = await enhancedDagTraversal({
         blockstore: this.helia.blockstore,
         signal: options?.signal,
         onProgress: options?.onProgress,
@@ -402,7 +383,6 @@ export class VerifiedFetch {
         offset,
         length,
         path,
-        response: tmpResponse,
         logger: this.helia.logger,
         contentTypeParser: this.contentTypeParser
       })
@@ -411,22 +391,14 @@ export class VerifiedFetch {
       const response = okRangeResponse(resource, byteRangeContext.getBody(), { byteRangeContext, log: this.log }, {
         redirected
       })
-      const contentType = tmpResponse.headers.get('content-type')
-      if (contentType != null) {
-        response.headers.set('content-type', contentType)
-      } else {
-        this.log('FIXME: content-type should be set')
-      }
 
-      // await setContentType({ bytes: firstChunk, path, response, contentTypeParser: this.contentTypeParser, log: this.log })
+      await setContentType({ bytes: firstChunk, path, response, contentTypeParser: this.contentTypeParser, log: this.log })
+
       setIpfsRoots(response, ipfsRoots)
 
       return response
     } catch (err: any) {
       options?.signal?.throwIfAborted()
-      // if (options?.signal?.aborted === true) {
-      //   throw new Error('aborted')
-      // }
       this.log.error('error streaming %c/%s', cid, path, err)
       if (byteRangeContext.isRangeRequest && err.code === 'ERR_INVALID_PARAMS') {
         return badRangeResponse(resource)
